+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use `k8smetadata` for labels and annotations instead of `apiextensions`.
+- Use `organization-operator` to import `Organization` CRD.
 
 ## [0.23.0] - 2022-03-02